requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
opencv-python>=4.9
numpy>=1.25
pandas>=2.0
scikit-learn>=1.3
albumentations>=1.4
pyyaml>=6.0
matplotlib>=3.8
tqdm>=4.66
tensorboard>=2.14